+++ examples/docs_snippets/docs_snippets/guides/dagster/dagster_pipes/databricks/resources.py
@@ -1,0 +1,17 @@
+import os
+
+from dagster_databricks import PipesDatabricksClient, WorkspaceClient
+
+import dagster as dg
+
+pipes_databricks_resource = PipesDatabricksClient(
+    client=WorkspaceClient(
+        host=os.environ["DATABRICKS_HOST"],
+        token=os.environ["DATABRICKS_TOKEN"],
+    )
+)
+
+
+@dg.definitions
+def resources():
+    return dg.Definitions(resources={"pipes_databricks": pipes_databricks_resource})
