+++ examples/docs_snippets/docs_snippets/guides/dagster/dagster_pipes/ecs/resources.py
@@ -1,0 +1,9 @@
+from dagster_aws.pipes import PipesECSClient
+
+import dagster as dg
+
+
+# start_resources
+@dg.definitions
+def resources():
+    return dg.Definitions(resources={"pipes_ecs_client": PipesECSClient()})
