+++ examples/docs_snippets/docs_snippets/guides/dagster/dagster_pipes/emr-serverless/resources.py
@@ -1,0 +1,10 @@
+from dagster_aws.pipes import PipesEMRServerlessClient
+
+import dagster as dg
+
+
+@dg.definitions
+def resources():
+    return dg.Definitions(
+        resources={"pipes_emr_serverless_client": PipesEMRServerlessClient()}
+    )
