+++ examples/docs_snippets/docs_snippets/guides/dagster/dagster_pipes/glue/dagster_code.py
@@ -1,7 +1,3 @@
-# start_asset_marker
-import os
-
-import boto3
 from dagster_aws.pipes import PipesGlueClient
 
 import dagster as dg
@@ -18,32 +14,3 @@ def glue_pipes_asset(
             "Arguments": {"some_parameter": "some_value"},
         },
     ).get_materialize_result()
-
-
-# end_asset_marker
-
-# start_definitions_marker
-
-import dagster as dg  # noqa
-from dagster_aws.pipes import PipesS3ContextInjector, PipesCloudWatchMessageReader
-
-
-@dg.definitions
-def resources():
-    return dg.Definitions(
-        resources={
-            "pipes_glue_client": PipesGlueClient(
-                client=boto3.client("glue"),
-                context_injector=PipesS3ContextInjector(
-                    client=boto3.client("s3"),
-                    bucket=dg.EnvVar("DAGSTER_GLUE_S3_CONTEXT_BUCKET"),
-                ),
-                message_reader=PipesCloudWatchMessageReader(
-                    client=boto3.client("logs")
-                ),
-            )
-        },
-    )
-
-
-# end_definitions_marker
