+++ examples/docs_snippets/docs_snippets/guides/dagster/dagster_pipes/lambda/dagster_code.py
@@ -1,4 +1,3 @@
-# start_asset_marker
 import boto3
 from dagster_aws.pipes import PipesLambdaClient
 
@@ -14,19 +13,3 @@ def lambda_pipes_asset(
         function_name="dagster_pipes_function",
         event={"some_parameter_value": 1},
     ).get_materialize_result()
-
-
-# end_asset_marker
-
-
-# start_definitions_marker
-@dg.definitions
-def resources():
-    return dg.Definitions(
-        resources={
-            "lambda_pipes_client": PipesLambdaClient(client=boto3.client("lambda"))
-        }
-    )
-
-
-# end_definitions_marker
