+++ examples/docs_snippets/docs_snippets/guides/dagster/dagster_pipes/lambda/resources.py
@@ -1,0 +1,13 @@
+import boto3
+from dagster_aws.pipes import PipesLambdaClient
+
+import dagster as dg
+
+
+@dg.definitions
+def resources():
+    return dg.Definitions(
+        resources={
+            "lambda_pipes_client": PipesLambdaClient(client=boto3.client("lambda"))
+        }
+    )
