+++ examples/docs_snippets/docs_snippets/guides/dagster/dagster_pipes/subprocess/part_2/step_2/dagster_code.py
@@ -11,10 +11,3 @@ def subprocess_asset(
     return pipes_subprocess_client.run(
         command=cmd, context=context
     ).get_materialize_result()
-
-
-@dg.definitions
-def resources():
-    return dg.Definitions(
-        resources={"pipes_subprocess_client": dg.PipesSubprocessClient()}
-    )
