+++ examples/docs_snippets/docs_snippets/guides/dagster/dagster_pipes/subprocess/with_asset_check/dagster_code.py
@@ -25,10 +25,3 @@ def no_empty_order_check(
         return dg.AssetCheckResult(passed=True)
 
     return dg.AssetCheckResult(passed=False)
-
-
-@dg.definitions
-def resources():
-    return dg.Definitions(
-        resources={"pipes_subprocess_client": dg.PipesSubprocessClient()}
-    )
