+++ examples/docs_snippets/docs_snippets/guides/non-python/pipes-asset.py
@@ -16,10 +16,3 @@ def tensorflow_model(
         context=context,
         extras={},
     ).get_materialize_result()
-
-
-@dg.definitions
-def resources():
-    return dg.Definitions(
-        resources={"pipes_subprocess_client": dg.PipesSubprocessClient()}
-    )
