+++ examples/docs_snippets/docs_snippets/guides/non-python/resources.py
@@ -1,0 +1,8 @@
+import dagster as dg
+
+
+@dg.definitions
+def resources():
+    return dg.Definitions(
+        resources={"pipes_subprocess_client": dg.PipesSubprocessClient()}
+    )
